change stories extension
